add jog command again

+++ arduino/nanokontrol_6axis/Preference.h
@@ -2,8 +2,8 @@
 #define _PREFERENCE_H
 
 typedef struct {
-	int dest;
-	int cur;
+	long dest;
+	long cur;
 } Motor;
 
 class Preference
